Updated VPlayer Music to 1.3.1
Added:
Added sound mute button.

Fixed:
Fixed issue on both players where when changing to master only input the video url bar decides to be
in a different location.
Fixed issue where videos would not load at all on the master only input.
Moved the position of the Old UI a little closer to the screen on both players.
Changed the Switch back to New UI text on the old UI to white on both players.