Fixes
Fixed video not auto playing the next video/ music on both players.
Removed further useless stuff from both players.
Changed the switch to new UI text on the old UI to white.
Fixed the Enter URL text not disapearing when the loading bar is present.
Fixed a bug with the master control where it was still allowing everyone to control the player (Not sure when this got introduced as a bug).

Known Issues:
The loading bar on the OLD UI broke, as the old UI and only there for debug reason's and is a legacy compoment this won't be fixed.