More Fixes
Fixed Issue with the infomation button not working on the VPlayer Music Only Player.
Fixed issue where sometimes the URL bar would not let you click it to enter a URL.
Updated changelog on the player.

This should be the finale build for version 1 of the player and no new updates should come out unto version 2 is ready unless we find bugs. This build will be deemed 1.3.3 and will be released on the 15th (Once we fully tested the build).